Removed .net framework project
