+++ csharp-advanced-topics/DelegatesInCsharp/ActionAndFuncDelegate/Entities/Person.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace ActionAndFuncDelegates.Entities
+namespace ActionAndFuncDelegate.Entities
 {
     public class Person
     {
