--- csharp-advanced-topics/DelegatesInCsharp/ActionAndFuncDelegates/ActionWithNoParameter.cs
@@ -1,20 +1,0 @@
-﻿using System;
-
-namespace ActionAndFuncDelegates
-{
-    internal class ActionWithNoParameter
-    {
-        static void Main(string[] args)
-        {
-            // Action delegate without parameters
-            Action printInformation = () => 
-            {
-                Console.WriteLine("It prints without parameter");
-            };
-
-            printInformation();
-
-            Console.ReadLine();
-        }
-    }
-}
